refactor: improve tests and types, add placeholder for queue and stack

+++ src/alg/search/binary_search.h
@@ -1,6 +1,9 @@
 #ifndef BINARY_H
 #define BINARY_H
 
-int binary_search(const int arr[], int start, int end, int target);
+#include <stdlib.h>
+#include <sys/types.h>
+
+int binary_search(const int arr[], ssize_t start, ssize_t end, int target);
 
 #endif
